refactor: extract JogWheel as reusable control component
Move JogWheel from DDJ-FLX4 device-specific location to shared
controls directory. The DDJ-FLX4 file now re-exports from the shared
location, so existing imports are unaffected. Add unit tests for
the shared component.
